make each transfomer as independent and move json specific to sub-module.
