+++ core/connectors/sdk/src/transforms/json/mod.rs
@@ -1,0 +1,8 @@
+// JSON-specific implementations for transforms
+pub mod add_fields;
+pub mod delete_fields;
+pub mod filter_fields;
+pub mod update_fields;
+
+#[cfg(test)]
+pub mod test_utils;
