+++ core/connectors/sdk/src/transforms/json/test_utils.rs
@@ -1,32 +1,9 @@
-/* Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-#[cfg(test)]
 use crate::{DecodedMessage, Payload, TopicMetadata};
-#[cfg(test)]
 use simd_json::OwnedValue;
-#[cfg(test)]
 use simd_json::prelude::{TypedScalarValue, ValueAsScalar};
-#[cfg(test)]
 use uuid;
 
 /// Helper function to create a test message with the given JSON payload
-#[cfg(test)]
 pub fn create_test_message(json: &str) -> DecodedMessage {
     let mut payload = json.to_string().into_bytes();
     let value = simd_json::to_owned_value(&mut payload).unwrap();
@@ -42,7 +19,6 @@ pub fn create_test_message(json: &str) -> DecodedMessage {
 }
 
 /// Helper function to create a non-JSON message with raw bytes
-#[cfg(test)]
 pub fn create_raw_test_message(bytes: Vec<u8>) -> DecodedMessage {
     DecodedMessage {
         id: None,
@@ -56,7 +32,6 @@ pub fn create_raw_test_message(bytes: Vec<u8>) -> DecodedMessage {
 }
 
 /// Helper function to create a topic metadata for testing
-#[cfg(test)]
 pub fn create_test_topic_metadata() -> TopicMetadata {
     TopicMetadata {
         stream: "test-stream".to_string(),
@@ -65,7 +40,6 @@ pub fn create_test_topic_metadata() -> TopicMetadata {
 }
 
 /// Helper function to extract the JSON object from a message
-#[cfg(test)]
 pub fn extract_json_object(msg: &DecodedMessage) -> Option<&simd_json::owned::Object> {
     if let Payload::Json(OwnedValue::Object(map)) = &msg.payload {
         Some(map)
@@ -75,7 +49,6 @@ pub fn extract_json_object(msg: &DecodedMessage) -> Option<&simd_json::owned::Ob
 }
 
 /// Helper function to assert that a JSON value is a number
-#[cfg(test)]
 pub fn assert_is_number(value: &OwnedValue, field_name: &str) {
     if !value.is_number() {
         panic!("{} should be a number", field_name);
@@ -83,7 +56,6 @@ pub fn assert_is_number(value: &OwnedValue, field_name: &str) {
 }
 
 /// Helper function to assert that a JSON value is a string
-#[cfg(test)]
 pub fn assert_is_string(value: &OwnedValue, field_name: &str) {
     if !value.is_str() {
         panic!("{} should be a string", field_name);
@@ -91,7 +63,6 @@ pub fn assert_is_string(value: &OwnedValue, field_name: &str) {
 }
 
 /// Helper function to assert that a JSON value is a string and validates as a UUID
-#[cfg(test)]
 pub fn assert_is_uuid(value: &OwnedValue, field_name: &str) {
     if !value.is_str() {
         panic!("{} should be a string", field_name);
