+++ core/connectors/sdk/src/transforms/update_fields.rs
@@ -16,75 +16,88 @@
  * under the License.
  */
 
-use regex::Regex;
+use super::{Transform, TransformType};
+use crate::{DecodedMessage, Error, Payload, TopicMetadata};
+use chrono::Utc;
 use serde::{Deserialize, Serialize};
 use simd_json::OwnedValue;
+use strum_macros::{Display, IntoStaticStr};
 
-use crate::{DecodedMessage, Error, Payload, TopicMetadata};
-
-use super::{
-    Transform, TransformType,
-    common::{Compilable, FieldValue, ValuePattern, compute_value},
-};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Field {
-    key: String,
-    value: FieldValue,
-    #[serde(default)]
-    condition: Option<UpdateCondition<String>>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct UpdateFieldsConfig {
-    fields: Vec<Field>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
+/// The value of a field, either static or computed at runtime
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum UpdateCondition<T = String> {
-    KeyExists,
-    KeyNotExists,
-    Matches(ValuePattern<T>),
-    DoesNotMatch(ValuePattern<T>),
+pub enum FieldValue {
+    Static(OwnedValue),
+    Computed(ComputedValue),
 }
 
-impl<T: Compilable> UpdateCondition<T> {
-    pub fn compile(self) -> Result<UpdateCondition<Regex>, Error> {
-        use UpdateCondition::*;
-        Ok(match self {
-            Matches(v) => Matches(v.compile()?),
-            DoesNotMatch(v) => DoesNotMatch(v.compile()?),
-            KeyExists => KeyExists,
-            KeyNotExists => KeyNotExists,
-        })
+/// Types of computed values that can be generated at runtime
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Serialize, Deserialize, Display, IntoStaticStr)]
+#[serde(rename_all = "snake_case")]
+pub enum ComputedValue {
+    #[strum(to_string = "date_time")]
+    DateTime,
+    #[strum(to_string = "timestamp_nanos")]
+    TimestampNanos,
+    #[strum(to_string = "timestamp_micros")]
+    TimestampMicros,
+    #[strum(to_string = "timestamp_millis")]
+    TimestampMillis,
+    #[strum(to_string = "timestamp_seconds")]
+    TimestampSeconds,
+    #[strum(to_string = "uuid_v4")]
+    UuidV4,
+    #[strum(to_string = "uuid_v7")]
+    UuidV7,
+}
+
+/// Computes a value based on the specified computed value type
+pub fn compute_value(kind: &ComputedValue) -> OwnedValue {
+    let now = Utc::now();
+    match kind {
+        ComputedValue::DateTime => now.to_rfc3339().into(),
+        ComputedValue::TimestampNanos => now.timestamp_nanos_opt().unwrap().into(),
+        ComputedValue::TimestampMicros => now.timestamp_micros().into(),
+        ComputedValue::TimestampMillis => now.timestamp_millis().into(),
+        ComputedValue::TimestampSeconds => now.timestamp().into(),
+        ComputedValue::UuidV4 => uuid::Uuid::new_v4().to_string().into(),
+        ComputedValue::UuidV7 => uuid::Uuid::now_v7().to_string().into(),
     }
 }
 
-struct CompiledField {
-    key: String,
-    value: FieldValue,
-    condition: Option<UpdateCondition<Regex>>,
+/// A field to be updated in messages
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Field {
+    pub key: String,
+    pub value: FieldValue,
+    #[serde(default)]
+    pub condition: Option<UpdateCondition>,
 }
 
+/// Configuration for the UpdateFields transform
+#[derive(Debug, Serialize, Deserialize)]
+pub struct UpdateFieldsConfig {
+    #[serde(default)]
+    pub fields: Vec<Field>,
+}
+
+/// Conditions that determine when a field should be updated
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum UpdateCondition {
+    Always,
+    KeyExists,
+    KeyNotExists,
+}
+
+/// Transform that updates fields in JSON messages based on conditions
 pub struct UpdateFields {
-    fields: Vec<CompiledField>,
+    pub fields: Vec<Field>,
 }
 
 impl UpdateFields {
-    pub fn new(cfg: UpdateFieldsConfig) -> Result<Self, Error> {
-        let mut out = Vec::with_capacity(cfg.fields.len());
-        for f in cfg.fields {
-            out.push(CompiledField {
-                key: f.key,
-                value: f.value,
-                condition: match f.condition {
-                    Some(c) => Some(c.compile()?),
-                    None => None,
-                },
-            });
-        }
-        Ok(Self { fields: out })
+    pub fn new(cfg: UpdateFieldsConfig) -> Self {
+        Self { fields: cfg.fields }
     }
 }
 
@@ -95,239 +108,16 @@ impl Transform for UpdateFields {
 
     fn transform(
         &self,
-        _meta: &TopicMetadata,
-        mut msg: DecodedMessage,
+        metadata: &TopicMetadata,
+        message: DecodedMessage,
     ) -> Result<Option<DecodedMessage>, Error> {
-        let Payload::Json(OwnedValue::Object(ref mut map)) = msg.payload else {
-            return Ok(Some(msg));
-        };
-
-        for f in &self.fields {
-            let present = map.contains_key(&f.key);
-            let pass = match &f.condition {
-                None => true,
-                Some(UpdateCondition::KeyExists) => present,
-                Some(UpdateCondition::KeyNotExists) => !present,
-                Some(UpdateCondition::Matches(pat)) => present && pat.matches(&map[&f.key]),
-                Some(UpdateCondition::DoesNotMatch(p)) => !present || !p.matches(&map[&f.key]),
-            };
-            if pass {
-                let val = match &f.value {
-                    FieldValue::Static(v) => v.clone(),
-                    FieldValue::Computed(c) => compute_value(c),
-                };
-                map.insert(f.key.clone(), val);
-            }
+        if self.fields.is_empty() {
+            return Ok(Some(message));
         }
 
-        Ok(Some(msg))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::transforms::common::{ComputedValue, FieldValue};
-    use crate::transforms::test_utils::{
-        assert_is_number, assert_is_uuid, create_raw_test_message, create_test_message,
-        create_test_topic_metadata, extract_json_object,
-    };
-    use simd_json::OwnedValue;
-
-    #[test]
-    fn test_basic_update() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![Field {
-                key: "status".to_string(),
-                value: FieldValue::Static(OwnedValue::from("updated")),
-                condition: None,
-            }],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 2);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "updated");
-    }
-
-    #[test]
-    fn test_condition_key_exists() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![
-                Field {
-                    key: "status".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("updated")),
-                    condition: Some(UpdateCondition::KeyExists),
-                },
-                Field {
-                    key: "missing_field".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("should not be added")),
-                    condition: Some(UpdateCondition::KeyExists),
-                },
-            ],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 2);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "updated");
-        assert!(!json_obj.contains_key("missing_field"));
-    }
-
-    #[test]
-    fn test_condition_key_not_exists() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![
-                Field {
-                    key: "status".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("should not update")),
-                    condition: Some(UpdateCondition::KeyNotExists),
-                },
-                Field {
-                    key: "created_at".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("2023-01-01")),
-                    condition: Some(UpdateCondition::KeyNotExists),
-                },
-            ],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 3);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "pending"); // Should remain unchanged
-        assert_eq!(json_obj["created_at"], "2023-01-01"); // Should be added
-    }
-
-    #[test]
-    fn test_condition_matches() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![
-                Field {
-                    key: "status".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("complete")),
-                    condition: Some(UpdateCondition::Matches(ValuePattern::Equals(
-                        OwnedValue::from("pending"),
-                    ))),
-                },
-                Field {
-                    key: "priority".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("high")),
-                    condition: Some(UpdateCondition::Matches(ValuePattern::Equals(
-                        OwnedValue::from("medium"),
-                    ))),
-                },
-            ],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending", "priority": "low"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 3);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "complete"); // Should be updated
-        assert_eq!(json_obj["priority"], "low"); // Should remain unchanged
-    }
-
-    #[test]
-    fn test_condition_does_not_match() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![
-                Field {
-                    key: "status".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("needs attention")),
-                    condition: Some(UpdateCondition::DoesNotMatch(ValuePattern::Equals(
-                        OwnedValue::from("complete"),
-                    ))),
-                },
-                Field {
-                    key: "priority".to_string(),
-                    value: FieldValue::Static(OwnedValue::from("high")),
-                    condition: Some(UpdateCondition::DoesNotMatch(ValuePattern::Equals(
-                        OwnedValue::from("low"),
-                    ))),
-                },
-            ],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending", "priority": "low"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 3);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "needs attention"); // Should be updated (not "complete")
-        assert_eq!(json_obj["priority"], "low"); // Should remain unchanged (is "low")
-    }
-
-    #[test]
-    fn test_computed_values() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![
-                Field {
-                    key: "updated_at".to_string(),
-                    value: FieldValue::Computed(ComputedValue::TimestampMillis),
-                    condition: None,
-                },
-                Field {
-                    key: "trace_id".to_string(),
-                    value: FieldValue::Computed(ComputedValue::UuidV4),
-                    condition: None,
-                },
-            ],
-        })
-        .unwrap();
-        let msg = create_test_message(r#"{"id": 1, "status": "pending"}"#);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        let json_obj = extract_json_object(&result).unwrap();
-        assert_eq!(json_obj.len(), 4);
-        assert_eq!(json_obj["id"], 1);
-        assert_eq!(json_obj["status"], "pending");
-        assert_is_number(&json_obj["updated_at"], "updated_at");
-        assert_is_uuid(&json_obj["trace_id"], "trace_id");
-    }
-
-    #[test]
-    fn test_non_json_payload() {
-        let transform = UpdateFields::new(UpdateFieldsConfig {
-            fields: vec![Field {
-                key: "status".to_string(),
-                value: FieldValue::Static(OwnedValue::from("updated")),
-                condition: None,
-            }],
-        })
-        .unwrap();
-        let msg = create_raw_test_message(vec![1, 2, 3, 4]);
-        let result = transform
-            .transform(&create_test_topic_metadata(), msg)
-            .unwrap()
-            .unwrap();
-        if let Payload::Raw(bytes) = &result.payload {
-            assert_eq!(bytes, &vec![1, 2, 3, 4]);
-        } else {
-            panic!("Expected Raw payload");
+        match &message.payload {
+            Payload::Json(_) => self.transform_json(metadata, message),
+            _ => Ok(Some(message)),
         }
     }
 }
